.safe_apply_tmp/pyproject.toml
[project]
name = "crypto-bot-pro"
version = "0.1.0"
description = "Installable crypto trading system (Mac/Windows) with safe execution and monitoring"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.36",
  "pydantic>=2.8",
  "httpx>=0.27",
  "websockets>=12.0",
  "orjson>=3.10",
  "SQLAlchemy>=2.0",
   "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "black>=24.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.briefcase]
project_name = "CryptoBotPro"
bundle = "com.cryptobotpro"
version = "0.1.0"
url = "https://localhost"
license = "Proprietary"
author = "CryptoBotPro"
author_email = "noreply@example.com"

[tool.briefcase.app.cryptobotpro]
formal_name = "CryptoBotPro"
description = "Crypto trading cockpit + engines (local Streamlit server)"
icon = "packaging/briefcase/icon"
sources = ["src/cryptobotpro_desktop"]
entry_point = "cryptobotpro_desktop.app:main"

# Keep this list minimal here; expand/sync with your real requirements as needed
requires = [
  "streamlit>=1.30",
  "pywebview>=5.0",
  "ccxt>=4.0",
  "pandas>=2.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
  "requests>=2.31",
]